+++ test/specs/enum.ts
@@ -2,67 +2,28 @@
 
 import 'mocha';
 
-import browserify from 'browserify';
 import { expect } from 'chai';
-import through from 'through';
 
-// @ts-ignore
-const transform = require('../dist').transform;
+import {
+  generateConfig,
+  resetConfig,
+  tagify,
+} from '../helpers/tagify';
 
-describe('String tags', function() {
+describe('Enum tags', function () {
   let output: string[] = [];
-  let config = {
-    env: {
-      CYPRESS_INCLUDE_TAGS: undefined,
-      CYPRESS_EXCLUDE_TAGS: undefined,
-    },
-  } as unknown as Cypress.PluginConfigOptions;
-
-  const tagify = (config: Cypress.PluginConfigOptions): Promise<string[]> => {
-    return new Promise((resolve, reject) => {
-      const options = {
-        typescript: require.resolve('typescript'),
-        extensions: ['.js', '.ts'],
-        plugin: [
-          ['tsify']
-        ],
-      };
-
-      try {
-        browserify(options)
-          .transform((fileName: string) => transform(fileName, config))
-          .add(__dirname + '/../cypress/integration/string.spec.ts')
-          .bundle()
-          .pipe(through(ondata, onend));
-
-        let data = ''
-        function ondata(d: string) { data += d }
-        function onend() {
-          const lines = data.split('\n')
-          const startline = lines.indexOf('// sample start') + 1;
-          const endline = lines.length - 3;
-
-          output = lines.slice(startline, endline);
-          resolve(output);
-        }
-      } catch (err) {
-        console.log('I have an error');
-        reject(err);
-      }
-    });
-  };
+  let config = generateConfig();
 
   beforeEach(() => {
-    delete config.env.CYPRESS_INCLUDE_TAGS;
-    delete config.env.CYPRESS_EXCLUDE_TAGS;
+    resetConfig(config);
   });
 
   describe('No tags provided', () => {
     before(async () => {
-      output = await tagify(config);
+      output = await tagify(config, 'enum');
     });
 
-    it('should output all tests without tags', function() {
+    it('should output all tests without tags', function () {
       expect(output).to.deep.equal([
         "describe('Run tests with no tags', () => {",
         "    it('I am a regular test', () => { });",
@@ -87,11 +48,11 @@ describe('String tags', function() {
 
   describe('Include tags provided', () => {
     before(async () => {
-      config.env.CYPRESS_INCLUDE_TAGS='wip';
-      output = await tagify(config);
+      config.env.CYPRESS_INCLUDE_TAGS = 'WIP';
+      output = await tagify(config, 'enum');
     });
 
-    it('should output all tests without tags', function() {
+    it('should output all tests without tags', function () {
       expect(output).to.deep.equal([
         "describe('Run tests with no tags', () => {",
         "    ;",
@@ -116,11 +77,11 @@ describe('String tags', function() {
 
   describe('Exclude tags provided', () => {
     before(async () => {
-      config.env.CYPRESS_EXCLUDE_TAGS='wip';
-      output = await tagify(config);
+      config.env.CYPRESS_EXCLUDE_TAGS = 'WIP';
+      output = await tagify(config, 'enum');
     });
 
-    it('should output all tests without tags', function() {
+    it('should output all tests without tags', function () {
       expect(output).to.deep.equal([
         "describe('Run tests with no tags', () => {",
         "    it('I am a regular test', () => { });",
@@ -140,12 +101,12 @@ describe('String tags', function() {
 
   describe('Include and exclude tags provided', () => {
     before(async () => {
-      config.env.CYPRESS_INCLUDE_TAGS='smoke,regression';
-      config.env.CYPRESS_EXCLUDE_TAGS='wip';
-      output = await tagify(config);
+      config.env.CYPRESS_INCLUDE_TAGS = 'SMOKE,REGRESSION';
+      config.env.CYPRESS_EXCLUDE_TAGS = 'WIP';
+      output = await tagify(config, 'enum');
     });
 
-    it('should output all tests without tags', function() {
+    it('should output all tests without tags', function () {
       expect(output).to.deep.equal([
         "describe('Run tests with no tags', () => {",
         "    ;",